# dermai-hospital/openapi.json
{"openapi":"3.1.0","info":{"title":"DermAI Hospital API","description":"Hospital-grade skin disease detection API with patient management, annotated image reports, RLHF feedback loop, and automated daily retraining.","version":"2.0.0"},"paths":{"/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","operationId":"register_auth_register_post","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string","title":"Username"}},{"name":"email","in":"query","required":true,"schema":{"type":"string","title":"Email"}},{"name":"password","in":"query","required":true,"schema":{"type":"string","title":"Password"}},{"name":"full_name","in":"query","required":false,"schema":{"type":"string","default":"","title":"Full Name"}},{"name":"department","in":"query","required":false,"schema":{"type":"string","default":"","title":"Department"}},{"name":"role","in":"query","required":false,"schema":{"type":"string","default":"doctor","title":"Role"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/patients/":{"post":{"tags":["Patients"],"summary":"Create Patient","operationId":"create_patient_patients__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"full_name","in":"query","required":true,"schema":{"type":"string","title":"Full Name"}},{"name":"gender","in":"query","required":false,"schema":{"type":"string","title":"Gender"}},{"name":"date_of_birth","in":"query","required":false,"schema":{"type":"string","title":"Date Of Birth"}},{"name":"contact_number","in":"query","required":false,"schema":{"type":"string","title":"Contact Number"}},{"name":"blood_group","in":"query","required":false,"schema":{"type":"string","title":"Blood Group"}},{"name":"medical_history","in":"query","required":false,"schema":{"type":"string","title":"Medical History"}}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Patients"],"summary":"List Patients","operationId":"list_patients_patients__get","security":[{"OAuth2PasswordBearer":[]}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/patients/{patient_id}":{"get":{"tags":["Patients"],"summary":"Get Patient","operationId":"get_patient_patients__patient_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"string","title":"Patient Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/predict/":{"post":{"tags":["Predictions"],"summary":"Predict Disease","operationId":"predict_disease_predict__post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_predict_disease_predict__post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/predict/{image_id}":{"get":{"tags":["Predictions"],"summary":"Get Prediction","operationId":"get_prediction_predict__image_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"image_id","in":"path","required":true,"schema":{"type":"string","title":"Image Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/feedback/":{"post":{"tags":["RLHF Feedback"],"summary":"Submit Feedback","description":"👍 vote=up  → model was correct, adds to positive RLHF pool\n👎 vote=down → provide correct_label; adds corrected sample to retrain pool","operationId":"submit_feedback_feedback__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/reports/{image_id}/image":{"get":{"tags":["Reports"],"summary":"Get Report Image","operationId":"get_report_image_reports__image_id__image_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"image_id","in":"path","required":true,"schema":{"type":"string","title":"Image Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reports/{image_id}/pdf":{"get":{"tags":["Reports"],"summary":"Get Report Pdf","operationId":"get_report_pdf_reports__image_id__pdf_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"image_id","in":"path","required":true,"schema":{"type":"string","title":"Image Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/dashboard":{"get":{"tags":["Admin"],"summary":"Admin Dashboard","operationId":"admin_dashboard_admin_dashboard_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/admin/model-versions":{"get":{"tags":["Admin"],"summary":"List Model Versions","operationId":"list_model_versions_admin_model_versions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/admin/trigger-retrain":{"post":{"tags":["Admin"],"summary":"Trigger Retrain","operationId":"trigger_retrain_admin_trigger_retrain_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/admin/audit-logs":{"get":{"tags":["Admin"],"summary":"Get Audit Logs","operationId":"get_audit_logs_admin_audit_logs_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["Health"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["Health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_login_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_auth_login_post"},"Body_predict_disease_predict__post":{"properties":{"patient_id":{"type":"string","title":"Patient Id"},"file":{"type":"string","format":"binary","title":"File"},"doctor_notes":{"type":"string","title":"Doctor Notes","default":""}},"type":"object","required":["patient_id","file"],"title":"Body_predict_disease_predict__post"},"FeedbackRequest":{"properties":{"image_id":{"type":"string","title":"Image Id"},"vote":{"type":"string","pattern":"^(up|down)$","title":"Vote"},"correct_label":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Correct Label"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["image_id","vote"],"title":"FeedbackRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/login"}}}}}}